+++ packages/serviceHelper/src/index.ts
@@ -1,0 +1,10 @@
+/*
+ * @Author: 邱狮杰
+ * @Date: 2023-06-04 09:58:45
+ * @LastEditTime: 2023-06-04 10:58:40
+ * @Description: 
+ * @FilePath: /memo/packages/serviceHelper/src/index.ts
+ */
+
+export type { QuickCompletionCRUDAssistantImpl, customTrigger, QuickCompletionCRUDAssistantConfig, ConfigureSuffix, TriggerResult, QuickCompletionCRUDAssistantResponse } from './quickCompletionCRUDAssistant/impl'
+export { unifyQuickCompletionCRUDAssistant, QuickCompletionCRUDAssistant } from './quickCompletionCRUDAssistant/index'