+++ packages/valuePolyfill/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @memo28/valuepolyfill
 
+## 1.0.2
+
+### Patch Changes
+
+- 修改 valuePolyfill 的 package.json 和新增 serviceHelper
+- Updated dependencies [9e3a714]
+- Updated dependencies [9e3a714]
+  - @memo28/utils@1.1.13
+
 ## 1.0.1
 
 ### Patch Changes